東京都発熱相談センターにおける相談件数: Rename components/cards/TokyoFeverConsultationCenterReportsNumberCard.vue to components/index/CardsReference/TokyoFeverConsultationCenterReportsNumber/Card.vue
