+++ libraries/repositories/FeverConsultationCenterRepository.ts
@@ -19,7 +19,7 @@ export class FeverConsultationCenterRepository
   /**
    * 使用箇所
    *
-   * 東京都発熱相談センターにおける相談件数 (components/cards/TokyoFeverConsultationCenterReportsNumberCard.vue)
+   * 東京都発熱相談センターにおける相談件数 (components/index/CardsReference/TokyoFeverConsultationCenterReportsNumber/Card.vue)
    */
   constructor() {
     super(Convert.toFeverConsultationCenter(JSON.stringify(rawData)))
